+++ Modules/Experience/Database/migrations/2025_04_25_171426_create_sessions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId( 'experience_id')->constrained('experineces_semesters')->onDelete('cascade');
             $table->foreignId( 'teacher_id')->constrained('users')->onDelete('cascade');
             $table->integer('status')->default(1);
-            $table->double('mark')->default(0);
+            $table->decimal('mark')->default(0);
             $table->timestamps();
         });
     }
