+++ Modules/Experience/Database/migrations/2025_06_03_084244_create_session_questions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('question');
             $table->enum('question_mode', ['fixed', 'dynamic']);
-            $table->double('questions_mark')->default(0);
+            $table->decimal('questions_mark')->default(0);
             $table->foreignId('session_id')->constrained('sessions')->onDelete('cascade');
             $table->timestamps();
         });
